Create Lab15_Animal_Invaders
This code contains all the functions needed to run the game. It may not work for your pot. I used a ~2cm pot and had a problem with linearity. If you update Player_Move() function with your numbers to work with your pot, the game should work fine. I used the default pins for the components.